third_party/blink/web_tests/platform/mac/external/wpt/css/css-variables/variable-presentation-attribute-expected.txt

This is a testharness.js-based test.
[FAIL] Testing 'alignment-baseline'.
  assert_equals: Default value. expected "baseline" but got "auto"
[FAIL] Testing 'baseline-shift'.
  assert_equals: Default value. expected "baseline" but got "0px"
[FAIL] Testing 'color-interpolation-filters'.
  assert_equals: Default value. expected "" but got "linearrgb"
[FAIL] Testing 'fill'.
  assert_equals: Default value. expected "black" but got "rgb(0, 0, 0)"
[FAIL] Testing 'flood-color'.
  assert_equals: Default value. expected "" but got "rgb(0, 0, 0)"
[FAIL] Testing 'font-family'.
  assert_equals: Default value. expected "Times New Roman" but got "Times"
[FAIL] Testing 'glyph-orientation-horizontal'.
  assert_equals: Default value. expected "0deg" but got ""
[FAIL] Testing 'glyph-orientation-vertical'.
  assert_equals: Default value. expected "auto" but got ""
[FAIL] Testing 'kerning'.
  assert_equals: Default value. expected "auto" but got ""
[FAIL] Testing 'lighting-color'.
  assert_equals: Default value. expected "" but got "rgb(255, 255, 255)"
[FAIL] Testing 'pointer-events'.
  assert_equals: Default value. expected "visiblePainted" but got "auto"
[FAIL] Testing 'stop-color'.
  assert_equals: Default value. expected "" but got "rgb(0, 0, 0)"
[FAIL] Testing 'stroke'.
  assert_equals: Default value. expected "" but got "none"
[FAIL] Testing 'writing-mode'.
  assert_equals: Default value. expected "lr-tb" but got "horizontal-tb"
Harness: the test ran to completion.
